+++ lib/routemaster/jobs/cache_and_sweep.rb
@@ -3,17 +3,24 @@ require 'routemaster/dirty/map'
 
 module Routemaster
   module Jobs
-    # Caches a URL using {Cache}, and sweeps the dirty map
-    # if sucessful.
+    # Caches a URL using {Cache} and sweeps the dirty map if successful.
+    # Busts the cache if the resource was deleted.
     class CacheAndSweep
       def perform(url)
         Dirty::Map.new.sweep_one(url) do
           begin
-            Routemaster::Cache.new.get(url)
+            cache.get(url)
           rescue Errors::ResourceNotFound
-            nil # nothing to cache
+            cache.bust(url)
+            true
           end
         end
+      end
+
+      private
+
+      def cache
+        @cache ||= Routemaster::Cache.new
       end
     end
   end