+++ lib/routemaster/middleware/dirty.rb
@@ -11,13 +11,14 @@ module Routemaster
     # The dirty map is passed as `:map` to the constructor and must respond to
     # `#mark` (like `Routemaster::Dirty::Map`).
     class Dirty
-      def initialize(app, dirty_map:nil)
+      def initialize(app, dirty_map: nil, **_)
         @app = app
         @map = dirty_map || Routemaster::Dirty::Map.new
       end
 
       def call(env)
         env['routemaster.dirty'] = dirty = []
+
         env.fetch('routemaster.payload', []).each do |event|
           next if event['type'] == 'noop'
           next unless @map.mark(event['url'])
@@ -28,6 +29,3 @@ module Routemaster
     end
   end
 end
-
-
-